Cleaned up yet more warnings...
git-svn-id: file:///fltk/svn/trunk@8498 ea41ed52-d2ee-0310-a9c1-e6b18d33e121

+++ src/drawtext.cxx
@@ -438,12 +438,12 @@ static void wrap(
 	// so add space to enhance the shortcut appearance 
 	// finally, the shortcut symbol should be centered for all platforms 
 	// whatever the font is so add a potential delta
-	int xdelta = (int) (getwidth(p,1)-getwidth(us,1))/2;
-	int ydelta =
+	float xdelta = (getwidth(p,1)-getwidth(us,1))/2.0;
+	float ydelta =
 #ifndef	_WIN32
-	    (getdescent())/2;
+	    (getdescent())/2.0;
 #else
-	(getdescent())/2 -1;
+	(getdescent())/2.0 - 1.0;
 #endif
 	add(0, us, us+1, x+xdelta, y+ascent+ydelta, getsize(),getsize(), ascent, spacing);
 }
